gitea_repos_master/simpler/simpler-register/requirements.txt
Faker==13.15.0
pymongo==4.1.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
tqdm==4.64.0
urllib3==1.26.10
